Move `RichDisplayURL` to the utils module
